Refactor rte_routines functions to static methods
