+++ db_routines.php
@@ -15,7 +15,6 @@ require_once 'libraries/common.inc.php';
  * Include all other files
  */
 require_once 'libraries/check_user_privileges.lib.php';
-require_once 'libraries/rte/rte_routines.lib.php';
 
 /**
  * Do the magic
