+++ libraries/classes/Rte/Events.php
@@ -104,7 +104,7 @@ class Events
 
             $item_query = self::getQueryFromRequest();
 
-            if (! count($errors)) { // set by PMA_RTN_getQueryFromRequest()
+            if (! count($errors)) { // set by PhpMyAdmin\Rte\Routines::getQueryFromRequest()
                 // Execute the created query
                 if (! empty($_REQUEST['editor_process_edit'])) {
                     // Backup the old trigger, in case something goes wrong
